Refactoring utils.py to have a new structure, more readable and maintainable
