+++ src/ci/input/testtoolconfig.json
@@ -1,0 +1,148 @@
+{
+    "wazuh_modules/syscollector":
+    [
+        {
+            "test_tool_name": "syscollector_test_tool",
+            "is_smoke_with_configuration": false,
+            "args":
+            [
+                "10"
+            ]
+        }
+    ],
+    "shared_modules/dbsync":
+    [
+        {
+            "test_tool_name": "dbsync_test_tool",
+            "is_smoke_with_configuration": true,
+            "args":
+            [
+                "-c config.json",
+                "-a snapshotsUpdate/insertData.json,snapshotsUpdate/updateWithSnapshot.json",
+                "-o ./output"
+            ]
+        },
+        {
+            "test_tool_name": "dbsync_test_tool",
+            "is_smoke_with_configuration": true,
+            "args":
+            [
+                "-c config.json",
+                "-a InsertionUpdateDeleteSelect/inputSyncRowInsert.json,InsertionUpdateDeleteSelect/inputSyncRowModified.json,InsertionUpdateDeleteSelect/deleteRows.json,InsertionUpdateDeleteSelect/inputSelectRows.json",
+                "-o ./output"
+            ]
+        },
+        {
+            "test_tool_name": "dbsync_test_tool",
+            "is_smoke_with_configuration": true,
+            "args":
+            [
+                "-c config.json",
+                "-a txnOperation/createTxn.json,txnOperation/inputSyncRowInsertTxn.json,txnOperation/pksGetDeletedRows.json,txnOperation/inputSyncRowModifiedTxn.json,txnOperation/closeTxn.json",
+                "-o ./output"
+            ]
+        },
+        {
+            "test_tool_name": "dbsync_test_tool",
+            "is_smoke_with_configuration": true,
+            "args":
+            [
+                "-c config.json",
+                "-a txnOperation/createTxn.json,txnOperation/inputSyncRowInsertTxn.json,txnOperation/fullyGetDeletedRows.json,txnOperation/inputSyncRowModifiedTxn.json,txnOperation/closeTxn.json",
+                "-o ./output"
+            ]
+        },
+        {
+            "test_tool_name": "dbsync_test_tool",
+            "is_smoke_with_configuration": true,
+            "args":
+            [
+                "-c config.json",
+                "-a triggerActions/insertDataProcesses.json,triggerActions/insertDataSocket.json,triggerActions/addTableRelationship.json,triggerActions/deleteRows.json",
+                "-o ./output"
+            ]
+        },
+        {
+            "test_tool_name": "dbsync_test_tool",
+            "is_smoke_with_configuration": true,
+            "args":
+            [
+                "-c config.json",
+                "-a triggerActions/insertDataProcesses.json,triggerActions/insertDataSocket.json,triggerActions/addTableRelationship.json,triggerActions/deleteRows.json",
+                "-o ./output"
+            ]
+        }
+    ],
+    "shared_modules/rsync":
+    [
+        {
+            "test_tool_name": "rsync_test_tool",
+            "is_smoke_with_configuration": false,
+            "args":
+            []
+        }
+    ],
+    "data_provider":
+    [
+        {
+            "test_tool_name": "sysinfo_test_tool",
+            "is_smoke_with_configuration": false,
+            "args":
+            []
+        }
+    ],
+    "syscheckd":
+    [
+        {
+            "test_tool_name": "fimdb_test_tool",
+            "smoke_tests_path": "src/db/smokeTests",
+            "is_smoke_with_configuration": true,
+            "output_folder": "./output/fileTransaction",
+            "args":
+            [
+                "-c config.json",
+                "-a FimDBTransaction/StartTransaction.json,FimDBTransaction/SyncTxnRows_1.json,FimDBTransaction/GetDeletedRows.json,FimDBTransaction/CountFiles.json,FimDBTransaction/StartTransaction.json,FimDBTransaction/SyncTxnRows_2.json,FimDBTransaction/GetDeletedRows.json,FimDBTransaction/CountFiles.json",
+                "-o ./output/fileTransaction"
+            ]
+        },
+        {
+            "test_tool_name": "fimdb_test_tool",
+            "smoke_tests_path": "src/db/smokeTests",
+            "is_smoke_with_configuration": true,
+            "output_folder": "./output/AtomicOperations",
+            "args":
+            [
+                "-c config.json",
+                "-a atomicFileOperations/SyncRow_1.json,atomicFileOperations/SyncRow_2.json,atomicFileOperations/CountFiles.json,atomicFileOperations/SyncRow_3.json,atomicFileOperations/DeleteFile.json,atomicFileOperations/CountFiles.json,atomicFileOperations/GetFile.json",
+                "-o ./output/AtomicOperations"
+            ]
+        }
+    ],
+    "winsyscheckd":
+    [
+        {
+            "test_tool_name": "fimdb_test_tool",
+            "smoke_tests_path": "src/db/smokeTests",
+            "is_smoke_with_configuration": true,
+            "output_folder": "./output/registryKeyTransaction",
+            "args":
+            [
+                "-c configWindows.json",
+                "-a FimDBTransaction/StartTransactionRegistryKey.json,FimDBTransaction/SyncTxnRowsRegistryKey_1.json,FimDBTransaction/GetDeletedRows.json,FimDBTransaction/StartTransactionRegistryKey.json,FimDBTransaction/SyncTxnRowsRegistryKey_2.json,FimDBTransaction/GetDeletedRows.json",
+                "-o ./output/registryKeyTransaction"
+            ]
+        },
+        {
+            "test_tool_name": "fimdb_test_tool",
+            "smoke_tests_path": "src/db/smokeTests",
+            "is_smoke_with_configuration": true,
+            "output_folder": "./output/registryDataTransaction",
+            "args":
+            [
+                "-c configWindows.json",
+                "-a FimDBTransaction/StartTransactionRegistryData.json,FimDBTransaction/SyncTxnRowsRegistryData_1.json,FimDBTransaction/GetDeletedRows.json,FimDBTransaction/StartTransactionRegistryData.json,FimDBTransaction/SyncTxnRowsRegistryData_2.json,FimDBTransaction/GetDeletedRows.json",
+                "-o ./output/registryDataTransaction"
+            ]
+        }
+    ]
+}
